Update security.md
In the chapter Validator-Validator Security in description: Take your validator infrastructure to the next level of security and availability: Composed of a cluster of signer nodes in place of the remote signe (https://docs.tendermint.com/master/nodes /remote-signer.html). This link results in an Error 404. The correct link to the material that is described is this: https://docs.tendermint.com/v0.34/tools/remote-signer-validation.htm